Moved duplicated gateway parameters to a trait.
